volume slider + mute button added
